requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=23.0.0
httpx>=0.24.0
openai>=1.0.0
qdrant-client>=1.7.0
PyYAML>=6.0.0
beautifulsoup4>=4.12.0
chardet>=5.2.0
requests>=2.31.0
tiktoken>=0.5.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.0.38
numpy>=1.26.0,<2.0.0

# State management dependencies
SQLAlchemy>=2.0.0
alembic>=1.12.0
appdirs>=1.4.4
aiosqlite>=0.19.0
greenlet>=3.0.0

# Git dependencies
GitPython>=3.1.40

# Jira dependencies
atlassian-python-api>=3.41.0 